requirements.txt
# Core
pdf2image>=1.17.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# APIs
httpx>=0.27.0
anthropic>=0.40.0
openai>=1.50.0

# Config & CLI
pyyaml>=6.0
rich>=13.0.0

# Web server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=12.0
